requirements.txt
asgiref==3.2.10
Django==3.1
django-autocomplete-light==3.5.1
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1

django-crispy-forms~=1.10.0